libstdc++: Adjust build of PCH files accounting configured host support.
This takes account of the overall configuration for host PCH support
when deciding if we should build the libstdc++ PCH files.

We now require both the support is configured and that we are hosted.
A non-fatal configure warning is given if the user attempts to
--disable-host-pch-support --enable-libstdcxx-pch since the latter
conflicts with the former (but does not prevent a useable libstdc++
library build).


libstdc++-v3/ChangeLog:

	* acinclude.m4: Account for configured host PCH support.
	* configure: Regenerate.
	* configure.ac: Act on --enable-host-pch-support.

+++ libstdc++-v3/acinclude.m4
@@ -3225,7 +3225,7 @@ AC_DEFUN([GLIBCXX_ENABLE_WCHAR_T], [
 ])
 
 
-dnl
+≈
 dnl Check to see if building and using a C++ precompiled header can be done.
 dnl
 dnl --enable-libstdcxx-pch=yes
@@ -3240,29 +3240,40 @@ dnl Substs:
 dnl  glibcxx_PCHFLAGS
 dnl
 AC_DEFUN([GLIBCXX_ENABLE_PCH], [
-  GLIBCXX_ENABLE(libstdcxx-pch,$1,,[build pre-compiled libstdc++ headers])
+  dnl This is only allowed if host support is enabled, and we are hosted.
+  if test "$1" = "yes" && test "$2" = "yes"; then
+    can_pch=yes
+  else
+    can_pch=no
+  fi
+  GLIBCXX_ENABLE(libstdcxx-pch,$can_pch,,[build pre-compiled libstdc++ headers])
   if test $enable_libstdcxx_pch = yes; then
-    AC_CACHE_CHECK([for compiler with PCH support],
-      [glibcxx_cv_prog_CXX_pch],
-      [ac_save_CXXFLAGS="$CXXFLAGS"
-       CXXFLAGS="$CXXFLAGS -Werror -Winvalid-pch -Wno-deprecated"
-       AC_LANG_SAVE
-       AC_LANG_CPLUSPLUS
-       echo '#include <math.h>' > conftest.h
-       if $CXX $CXXFLAGS $CPPFLAGS -x c++-header conftest.h \
+    if test "$2" != "yes"; then
+      glibcxx_cv_prog_CXX_pch=no
+      AC_MSG_WARN([PCH headers cannot be built since host PCH is disabled])
+    else
+      AC_CACHE_CHECK([for compiler with PCH support],
+        [glibcxx_cv_prog_CXX_pch],
+        [ac_save_CXXFLAGS="$CXXFLAGS"
+         CXXFLAGS="$CXXFLAGS -Werror -Winvalid-pch -Wno-deprecated"
+         AC_LANG_SAVE
+         AC_LANG_CPLUSPLUS
+         echo '#include <math.h>' > conftest.h
+         if $CXX $CXXFLAGS $CPPFLAGS -x c++-header conftest.h \
 			  -o conftest.h.gch 1>&5 2>&1 &&
 		echo '#error "pch failed"' > conftest.h &&
 	  echo '#include "conftest.h"' > conftest.cc &&
 	       $CXX -c $CXXFLAGS $CPPFLAGS conftest.cc 1>&5 2>&1 ;
-       then
-	 glibcxx_cv_prog_CXX_pch=yes
-       else
-	 glibcxx_cv_prog_CXX_pch=no
-       fi
-       rm -f conftest*
-       CXXFLAGS=$ac_save_CXXFLAGS
-       AC_LANG_RESTORE
-      ])
+         then
+	   glibcxx_cv_prog_CXX_pch=yes
+         else
+	   glibcxx_cv_prog_CXX_pch=no
+         fi
+         rm -f conftest*
+         CXXFLAGS=$ac_save_CXXFLAGS
+         AC_LANG_RESTORE
+        ])
+    fi
     enable_libstdcxx_pch=$glibcxx_cv_prog_CXX_pch
   fi
